Prefilter classes and functions in TypeSolver.solve.
Pass in the unknowns pytd and the builtins separately to the
TypeSolver and partition the classes and functions in the former
into unknown, partial, and complete subsets before generating
equations for solving. This also lets us pass the pytd with
NamedTypes changed to ClassTypes back to convert_pytd, which ensures
that NamedTypes do not show up in the solution.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=96459808