Fix a bug in CallTracer._call_traces_to_functions.
This method was getting the argument names for a function from its
first signature, which was causing two problems:

* When the signature lengths are different, pytype can
mysteriously crash. Suppose we have the following signatures:
  def f(x: str) -> float
  def f(x: int, y: bool) -> long
and want to analyzing f(0, True). Because pytype grabs the first
signature, the second argument is accidentally dropped from the
call trace, turning it into f(x: int) -> ?, which matches neither
signature, so a FlawedQuery is raised.

* When the argument names are different, pytype can produce bad
results. Suppose we have:
  def f(arg1: str) -> float
  def f(arg2: int) -> long
and want to analyze
  def g(x):
    return f(x)
pytype grabs the first signature and therefore assumes the argument
name to f is "arg1", leading to the incorrect inferred signature
  def g(x: str) -> float
for g.

This CL anonymizes the names of positional args to resolve these problems.
(So, for example, f(0, True) turns into f(_0: int, _1: bool).)

--
MOS_MIGRATED_REVID=134604517