Replace the classes in typing.py with abstract.py equivalents.
Right now typing.py duplicates a lot of classes (e.g., there exists
both an abstract.Union and a typing.Union; the type List[int] is
represented as an abstract.ParameterizedClass when converted from pyi
but as a typing.List value when converted from an inline annotation).

With this change, the typing.py classes build their abstract.py
equivalents when indexed into. For example:

Old behavior:
  from typing import Union  # Creates a typing.Union object
  def f(x: Union[str, unicode]): ...  # Calls the getitem slot on typing.Union,
                                      # which returns a new typing.Union object
                                      # for use as the annotation for x.

New behavior:
  from typing import Union  # Creates a typing.Union object
  def f(x: Union[str, unicode]): ...  # Calls getitem on typing.Union, which
                                      # returns an abstract.Union object for
                                      # use as the annotation.

In cases in which the bare typing.py object is used as an annotation,
VirtualMachine._process_one_annotation replaces it with the corresponding
abstract.py class:
  from typing import List  # Creates a typing.Container(List) object
  def f(x: List): ...  # _process_one_annotation replaces
                       # typing.Container(List) with
                       # abstract.ParameterizedClass(List).

To keep various tests passing, this change also
* moves the logic for setting a ParameterizedClass's module from convert.py
  to ParameterizedClass.__init__,
* changes the matcher to allow formal types on the left,
* tweaks Instance.compatible_with to handle container instances with
  uninitialized type parameters.

--
MOS_MIGRATED_REVID=140524544