Improve the definitions of __builtin__.generator and typing.Generator.
Changes a generator to have three type parameters and its send method to
be more precise
(https://github.com/python/typeshed/blob/master/stdli[]/typing.pyi).

Since it is common to specify only one parameter for a generator, this
change also tweaks convert.py and typing.py to support specifying only
some of the parameters to a generic and filling in unsolvable for the rest.

--
MOS_MIGRATED_REVID=140772693