Adds support for passing a view to to_type and get_instance_type.
When an unsolvable is among the possible values for a parameter
that fails type matching, bad_matches prints "Any" for the parameter,
hiding the bad value. This change adds a "view" argument to to_type
and get_instance_type so that bad_matches can specify that only the
bad value of a parameter be included when converting to a pytd type.

--
PiperOrigin-RevId: 141459179
MOS_MIGRATED_REVID=141459179