Make __builtin__.tuple homogeneous.
Changes both parsers to check for __builtin__.tuple in addition to
typing.Tuple when deciding whether to make a HomogeneousContainerType.

The definition of `dictionary-itemiterator`.next() previously had to
use "tuple[K or V]" because "tuple[K, V]" was parsed into a
heterogeneous tuple (causing testDictIterators in test_builtins2 to fail);
it's now safe to switch to the slightly better definition.

--
PiperOrigin-RevId: 142274634
MOS_MIGRATED_REVID=142274634