Fix some tuple definitions in our pytd files.
Replaces tuple[X] with tuple[X, ...], tuple[?] with tuple,
  def foo() -> tuple[X]  # tuple[X, X]
with
  def foo() -> tuple[X, X],
and tuple[X or Y] with tuple[X, Y] when it is obvious that the latter
was intended. Also collapses a couple of
"list[T] or set[T] or tuple[T] or ..."-type definitions to "Iterable[T]"
and deletes tuple[nothing] when it appears in a union of homogeneous tuples.

--
PiperOrigin-RevId: 142486324
MOS_MIGRATED_REVID=142486324