Get rid of Variable.name.
The "name" of a (pytype) Variable is a relict, and has provided little value.
Almost all of our Variables are abstract items that refer to e.g. a position
on a stack, a temporary result, a subexpression of an if-splitting condition
or a dummy value. As such, they're difficult to name, and the names given
to them by existing code were mostly useless, if not misleading.

Variable names were also a source of bugs: The incorrect names would tend to
leak into analysis results or error messages.

When inspecting a Variable for debugging, it's much more insightful to look at
[x.name for x in variable.data].

We will also add a new logging statement to cfg.Program.NewVariable that fires
whenever a new Variable is introduced, which makes it much easier and reliable
to discover the origins of a Variable in the logs.

--
PiperOrigin-RevId: 143456613
MOS_MIGRATED_REVID=143456613