Be smarter about what top-level defs we ignore in pyi generation.
We ignore builtin names only if the code we're analyzing hasn't
redefined them.

Names from typing.pytd are ignored so that our pyi files aren't
littered with
  List = ...  # type: type
and similar.

This change causes the inferred pyi for, e.g.,
  def f(x: Any): pass
to change from
  from typing import Any
  def f(x: Any): ...
to
  from typing import Any
  def f(x): ...
which breaks tests because the unnecessary import is removed from
the expected pytd in assertTypesMatchPytd, so we tweak PrintVisitor
to omit the import in such cases.

--
PiperOrigin-RevId: 143589706
MOS_MIGRATED_REVID=143589706