Improve our handling of TooComplexError.
Since deep_variable_product can be called through bad_matches,
it was possible for _check_return to raise a TooComplexError
that we don't catch. (I don't have a test case for this bug
because I only managed to trigger it through an experimental
change to abstract.Tuple.cls.)

abstract.get_views is now the only place where
deep_variable_product is directly called, so we can catch the
exception as soon as it is raised.

--
PiperOrigin-RevId: 144676410
MOS_MIGRATED_REVID=144676410