Speed up the data flow checking.
* Remove a call to sorted(): 
For a mildly complex file visitors.py of pytype sorted is called 6 million times and costs about ~9 seconds.
This cost does not need to be payed, as we can as well just cache one list and use it instead, no special order is necessary.

* Early search exit:
After proving that a path without conditions exist or two path exist, which do not share a node containing a condition, the search can be aborted.

--
PiperOrigin-RevId: 145542278
MOS_MIGRATED_REVID=145542278