Improve long-tail performance of if-splitting.
This fixes a bug in how a chain of conditions is added to goal, and also adds
an approximation step that removes conditions that don't seem related to the
current goal.

Consider the following code:

def f(x: Union[A, B], y: int): # n1
  z = true_or_false()          # n2
  if z:                        # n3
    return (x, y)              # n4

When we solve for the return value (say, for (A, int)), we try to trace a path
back to the defintion of x and y (i.e., the start of the function). On that
path, we encounter a condition z=True.

Before:
  We'd add the new goal z=True, introduced by the condition, to our list of
  goals, and re-run the solver at n3. Since we weren't able to expand any
  goals, this is a second parallel solver run, multiplying the solving time by
  two. In other words, the behavior is exponential in the number of nested if
  statements.

After:
  We see, but do not add, the goal z=True, since it's fulfilled by assignments
  on a path to n1. (That's an approximation, but one that holds up against all
  the test cases we have.)
  We then jump to n1, remove the goals x=A and y=int and finish solving.

--
PiperOrigin-RevId: 147473789
MOS_MIGRATED_REVID=147473789