Remove wrong __slots__ statements in pytd.py and make the others have an effect.
The once removed were wrong because they did not declare instance variables. This is not easy to fix while keeping the __slots__ statetement as children of NamedTuple are not allowed non empty __slots__ statements.

As node.Node returned a class without a __slots__ statement the childrens __slots__ statement were effectless.
"When inheriting from a class without __slots__, the __dict__ attribute of that class will always be accessible, so a __slots__ definition in the subclass is meaningless."
from
https://docs.python.org/2/reference/datamodel.html

--
PiperOrigin-RevId: 149277345
MOS_MIGRATED_REVID=149277345