Speedup _deep_values_list_product.
Replace a tuple with a set. The set is faster, as we pay constant time for the lookups in line 126 instead of linear time.

Profiling
This was tested for
video/youtube/src/python/config/shmoovies.py
which is a configuration file that is known to be very slow.
Before this change 121 seconds were spend in _deep_values_list_product, after this change it is 6s, in percent that is 95% of the load removed.
In total before this change profiling the call took ~163 seconds, with this change it is ~50s for a speedup factor of roughly 3x.
 
[]
shows that all calls sites were updated with set() instead of (). 

I suspect that we could as well use a seen.update to save on the number of sets, but haven't investigated this and changing it could change behavior.

Command
./blaze-bin/third_party/py/pytype/pytype  --check --abort-on-complex --no-report-errors --output="-" --imports_info=blaze-out/gcc-4.X.Y-crosstool-v18-hybrid-grtev4-k8-opt/bin/video/youtube/src/python/config/pytype-aspect/shmoovies@pytype@check/imports_info --python_exe="third_party/py/pytype/google/python2.7 -T" --module-name=[].video.youtube.src.python.config.shmoovies --quick --precompiled-builtins=blaze-out/gcc-4.X.Y-crosstool-v18-hybrid-grtev4-k8-opt/bin/devtools/python/blaze/pytype/release/precompiled_builtins.out video/youtube/src/python/config/shmoovies.py --output="test.pyi" --profile='profile2'

--
PiperOrigin-RevId: 149277428
MOS_MIGRATED_REVID=149277428