Manage a set of ancestors instead of finding them on the fly.
cfg.py:CanHaveCombination dominates the runtime not spend ingesting PYI files.
A lot of the runtime of that function is spend just finding the ancestors of nodes to compare the bindings in them with a list of supplied bindings (goals).

This CL aims to reduce the lookup time by guaranteeing that reachable_subset is always up to date. 
With this guarantee the work of finding the ancestors only needs to be done during CFG modification.
The chosen approach updates the self.ancestors of a child if a new connection is added at the parent of the child.

Profiling: CanHaveCombination goes from being ~50% of the total runtime, for files containing many dictionaries (configurations), down to less than 0.5%.
The test file took 240seconds, of which 144s were in CanHaveCombination. This did go down to 0.3s in CanHaveCombination and 1.5s in _Update_Reachable_Backwards. Total runtime is now ~80s.

--
PiperOrigin-RevId: 149928477
MOS_MIGRATED_REVID=149928477