Set the filename appropriately when analyzing __builtin__.py.
This bug's an odd one. For tests, the filename is None, so the
  op.code.co_filename == self.vm.filename
check in apply_type_comment succeeds when we're preloading the
builtins from __builtin__.py. Thus, we end up analyzing the
builtins file but applying type comments from the test code.
This is usually harmless, but __builtin__.py has a docstring
that ends at line 4, so if the fourth line of the test code
happens to have a type comment that uses names which __builtin__.py
doesn't import, then we get baffling errors.

(The test in test_typevar.py will be used for testing that we
report errors for "# type: T".)

--
PiperOrigin-RevId: 151740791
MOS_MIGRATED_REVID=151740791