Don't cache values that use non-root nodes.
Remove the "node" argument to _constant_to_value so that we can't accidentally
hide a node in a cached value. Instead, _constant_to_value takes a get_node
function that returns the current node. Then constant_to_value can define
get_node to set a need_node flag so that we know to return the value without
caching it.

--
PiperOrigin-RevId: 152023011
MOS_MIGRATED_REVID=152023011