Only allow using a TypeVar from __builtin__.pytd if it is imported.
Logs a name error if byte_LOAD_NAME loads a type parameter from
builtins (since that means the type parameter wasn't in locals or
globals, and thus wasn't imported). Importing a TypeVar is still
allowed - the names are now prefixed with underscores, which should
be sufficient warning that they're not meant for public use.

--
PiperOrigin-RevId: 152192051
MOS_MIGRATED_REVID=152192051