Delay logging key_error until we're sure it's an error.
Changes Dict.getitem_slot to raise a DictKeyMissing error instead of
logging key error so that when dict is one of multiple variable
bindings, the error is not logged prematurely.

A side effect of this change is that we sometimes log one of the
FailedFunctionCall errors (e.g., wrong-arg-types) when we previously
would have logged unsupported-operands.

--
PiperOrigin-RevId: 152570626
MOS_MIGRATED_REVID=152570626