Put the current line number in the names of cfg nodes.
Sets the names of nodes to the current line number in vm.py and state.py.
Also adds the line number to Opcode.__str__ to make it easier to match
nodes to opcodes when comparing the debug output to the cfg, since the node
name can't be too long (or we run into problems with rendering the graph).

--
PiperOrigin-RevId: 152734762
MOS_MIGRATED_REVID=152734762