Support tuples and unions of type parameters in class definitions in pyi.
We previously handled definitions like
  class Foo(List[Tuple[_K, _V]]): ...
in a silly incorrect way: since Tuple[_K, _V] is formal, we'd record
"_T" as an alias to "tuple", which meant that we never found any
substitutions for it.

Instead, we'll instantiate tuples of type parameters using
TypeParameterInstance. This requires adding a "container" argument to
AtomicAbstractValue.instantiate so that each TypeParameterInstance knows
what instance it belongs to, as well as tweaking matcher.py.

--
PiperOrigin-RevId: 152865805
MOS_MIGRATED_REVID=152865805