Add an abstract.List that remembers the initial list contents.
Similar to abstract.Tuple, abstract.List stores the contents
that build_list gives it. Since lists are mutable, abstract.List
has a could_contain_anything attribute that it sets to True
when merge_type_parameter is called, to indicate that the original
contents are no longer usable.

This change is motivated by typing.Callable's need to know the
positions of the types in the list of argument types that it gets.

--
PiperOrigin-RevId: 153203247
MOS_MIGRATED_REVID=153203247