Don't create unknowns for return values with unsubstituted params.
Instead of creating an unknown, we set params without substitutions
to Empty. This fixes an odd bug where, when a decorator returns a
type with an unsubstituted type parameter, a decorated method on a
class turns into an unknown, causing us to trace its calls and use the
call traces to generate a (possibly silly) signature.

--
PiperOrigin-RevId: 153471524
MOS_MIGRATED_REVID=153471524