Remove restriction that HasSlots can only be added to SimpleAbstractValue.
The reason for the SimpleAbstractValue check in HasSlots.init_mixin was that
set_slot was using get_instance_attribute, which requires a
SimpleAbstractValue. The use of get_instance_attribute instead of
get_attribute was due to the latter calling get_special_attribute (which
would return the slot we're in the process of setting, introducing an
infinite loop). This issue was resolved by simply moving the get_attribute
call to before the slot is set.

--
PiperOrigin-RevId: 153616128
MOS_MIGRATED_REVID=153616128