Add functionality to generate a pickled ast representation during inference.
This will solve some runtime overhead, as the dependencies are already loaded during inference and can be reused.
This also adds an option to read the generated pyi file and compare it to the ast generated for pickling. This option allows for easier debugging of issues in the pickling code, as failures can be caught at saving time instead of loading time.

--
PiperOrigin-RevId: 153685750
MOS_MIGRATED_REVID=153685750