Implement matching of return types for PyTDFunction.
Adds some of the infrastructure needed to match functions
against abstract.Callable and does one of the simplest pieces
(matching the return type for a PyTDFunction).

We still need to do matching for arguments, InterpreterFunction,
and BoundFunction, as well as print something better than a
plain "Callable" in error messages.

--
PiperOrigin-RevId: 153790077
MOS_MIGRATED_REVID=153790077