Support matching InterpreterFunction against Callable.
Completes the TODO in matcher.py for InterpreterFunction vs Callable.

Stops allowing specifying the type of "self" when matching a BoundFunction
against a Callable, since the function won't actually accept a "self" argument.

Does matching even when an annotation doesn't exist, so that type parameters
are filled in properly.

--
PiperOrigin-RevId: 154121665
MOS_MIGRATED_REVID=154121665