Remove the requirement that pytd.CallableType has to have at least two params.
I'm not sure why I thought putting in a "nothing" placeholder when there are no
arguments was a good idea.

Simplifies some code by removing this requirement.
Stops (incorrectly) simplifying TupleType and CallableType in
SimplifyContainers; the parameter length is meaningful even when all of the
parameters are Any.

--
PiperOrigin-RevId: 154209434
MOS_MIGRATED_REVID=154209434