Remove the "subst" argument to bad_matches().
My previous change introduced a bug by overwriting "subst" with None
when a bad match was found. (Our tests didn't catch this because the
error only shows up when we iterate through the views in a particular
order.) Since bad_matches is never called with the "subst" argument, we can
just remove it.

--
PiperOrigin-RevId: 154575843
MOS_MIGRATED_REVID=154575843