Use LookupExternalTypes during unpickle.
Change unpickling code to use LookupExternalTypes in all cases instead of special casing all the special cases we know about.
This is needed as we discover more and more special cases, in which the AST is structurally modified during dependency resolution in ways which can not be predicted during inference. This includes
* The type of external dependencies changing between inference and loading of the inferred ast.
* External modules using a __getattr__ -> Any rule which was not known at inference time.

The change will have some performance impact, as one iteration over the AST will be needed. Due to the change in the LookupExternalTypes Visitor this iteration will in most cases not structurally modify the AST.

--
PiperOrigin-RevId: 154832228
MOS_MIGRATED_REVID=154832228