Rollback of commit c694317e225238b030365fcba4f3f2661878ca68.
*** Reason for rollback ***

Broke pytype

*** Original change description ***

Use LookupExternalTypes during unpickle.

Change unpickling code to use LookupExternalTypes in all cases instead of special casing all the special cases we know about.
This is needed as we discover more and more special cases, in which the AST is structurally modified during dependency resolution in ways which can not be predicted during inference. This includes
* The type of external dependencies changing between inference and loading of the inferred ast.
* External modules using a __getattr__...

***

--
PiperOrigin-RevId: 154839974
MOS_MIGRATED_REVID=154839974