Support `x in iter(container)`.
Support CMP_IN/CMP_NOT_IN on an iterable without a __contains__ method. When
evaluating the `in` operator for an object without a __contains__ attribute,
the interpreter gets the object's iterator and does comparisons against the
results of calling next(). We simply attempt to get the iterator and return an
abstract boolean value upon success (i.e., we skip the next() calls).

--
PiperOrigin-RevId: 154889660
MOS_MIGRATED_REVID=154889660