Fix serialize_ast handling for __init__.py files and relative imports.
Pytype expects __init__.py files to be inferred with the module name suffix .__init__ this marker is needed for resolution of relative imports.
During reference resolution __init__.py files are expected to have to have the name of the module, e.g. "foo" instead of "foo/__init__".
This name change was resolved in the EnsureName method so far, to save those AST visits, the name should be changed before serialization.

--
PiperOrigin-RevId: 155191020
MOS_MIGRATED_REVID=155191020