Generate a traceback string from the frames passed to Error.with_stack.
Generates a traceback string with information about up to three of the frames
in the stack. Saves the string as a "_traceback" attribute (unused for now, but
it will later be added to Error.__str__ and ErrorLogBase.print_to_csv_file).

I previously thought we would need to save the method name as a new attribute
on the frame, but it turns out that frame.current_opcode.code.co_name is the
method name.

--
PiperOrigin-RevId: 155917612
MOS_MIGRATED_REVID=155917612