Fix caching in InterpreterFunction.call(), and a bug in BuildClass.call().
Fixes caching of locals in InterpreterFunction.call(). This change exposes
a bug in our implementation of Python 3's LOAD_BUILD_CLASS: when a class 'X'
is built, a new locals dictionary is created for X's members, but the
'f_locals' attribute of InterpreterFunction(X) still points to the old
locals (which happen to be the globals when X is a top-level class), so when
InterpreterFunction(X) is analyzed by CallTracer.maybe_analyze_method, all of
X's members are added to the globals. We fix this bug by having
BuildClass.call() set the interpreter function's 'f_locals' to the new locals.

--
PiperOrigin-RevId: 155927074
MOS_MIGRATED_REVID=155927074