Stop specifying the filename in tests in test_type_comments.
It used to be the case that omitting the filename in tests could cause
spurious name errors due to type comments being attributed to the wrong file.
This was fixed a while ago, so we can stop specifying 'filename="test.py"'.
There was a test (testTypeCommentUsesFilename) that depended on this 'feature',
so that test has been removed entirely.

--
PiperOrigin-RevId: 156323834
MOS_MIGRATED_REVID=156323834