Create an index of ClassType nodes to avoid AST visits.
Create a list of all ClassType nodes in an AST during serialization. This list is used during deserialization to avoid a visit to the whole AST, instead only the ClassType nodes are visited to fill the .cls pointers.
If one of the ClassTypes is found to have changed its type during the visit, which can happen if serialization and deserialization use different ASTs for the dependencies, the whole AST is visited. This visit is necessary because we can't exchange a node without visiting the parent.

--
PiperOrigin-RevId: 157574451
MOS_MIGRATED_REVID=157574451