Directly fill in subst when matching an unknown.
When the abstract matcher has an AMBIGUOUS_OR_EMPTY value on the left, it
was previously instantiating the type on the right and continuing with
matching in order to fill in subst. This caused problems because, e.g.,
Tuple[str or unicode, str or unicode] doesn't match Tuple[AnyStr, AnyStr].
Instead, use annotations_util's get_type_parameters method to find the
type parameters that need to be added to subst without doing further matching.


Also includes the following changes:

Handle super(...).__new__/__init__ more carefully.

When object.__new__ or object.__init__ is called via super(), the method
accepts and ignores (any number of) extra arguments iff the other one of the
methods is overridden. For example, the following is legal:

class Foo(object):
  def __new__(cls):
    return super(Foo, cls).__new__(cls, 1, 2, 3, 4, 5, 6)  # ok
  def __init__(self):
    pass

but the following is not:

class Foo(object):
  def __new__(cls):
    return super(Foo, cls).__new__(cls, 1, 2, 3, 4, 5, 6)  # error!

This change adds a special_builtins.Object class that uses
get_special_attribute to return either the normal pytd definition of
object.__new__/__init__ or a NativeFunction implementation that swallows
extra arguments.

Also adds a skipped test for generating and reloading a namedtuple definition.
This doesn't work yet because our tuple definition overrides object.__init__.
We need to either switch tuple over to __new__ and fix everything that the
switch breaks (which would be the principled thing to do), or add an __init__
definition to NamedTupleBuilder (incorrect, but easy and maybe good enough).

--
PiperOrigin-RevId: 157652139
MOS_MIGRATED_REVID=157652139