Generate type parameters for __new__.
Previously, when a class defined its own __new__ method that called
object.__new__ via super(), it ended up with a definition like this:

class X(object):
  def __new__(cls) -> X

The issue is that if X is subclassed in another module, that
subclass can't be instantiated, since __new__ only produces instances
of X. This is particularly troublesome for namedtuple because

class X(namedtuple("_X", ...)):
  def __new__(cls, ...): ...

is a common pattern, and a namedtuple instance is a PyTDClass, so
X.__new__ behaves as if _X.__new__ is from another module.

This is fixed by inserting bounded type parameters into the generated
pyi for __new__. (The namedtuple pattern still doesn't work, for reasons
unrelated to this change.)


Also includes the following changes:

Add hasattr() to special_builtins, to support if-splitting.

--
Adds overlay for abc module, implements abc.abstractmethod decorator. Adds a
check to avoid analyzing abstract methods unless they are called using super.

--
PiperOrigin-RevId: 158184267
MOS_MIGRATED_REVID=158184267