Use parse_string in collections_overlay.NamedTupleBuilder.
We can use parse_string to create the namedtuple instance from a string as
long as we fill in the class types afterwards.

Some of the namedtuple tests are now a little silly, since they essentially
check that the generated ast equals itself, so most of their remaining value
is in verifying the absence of crashes or unexpected errors.

--
PiperOrigin-RevId: 158447678
MOS_MIGRATED_REVID=158447678