Split the "seen" arg to the analyze_x methods into two CallTracer attributes.
This argument was doing double-duty as a set of seen instances and a set of
seen functions, so split it into two private attributes on CallTracer.


Also includes the following changes:

Adds method to ensure __init__ and __new__ calls always succeed.

If an __init__ or __new__ call fails, report the error and call it again with
made-up arguments. For simplicity, it chooses the first possible function
signature available.

Adds a test for this behavior.

This should eventually be expanded to all functions, not just __new__ and __init__. See related bug for discussion about why this can't be expanded already. (Short version: no support for keyword-only arguments, tests assume calls will fail.)

--
PiperOrigin-RevId: 158564577
MOS_MIGRATED_REVID=158564577