Handle unbound super instances better.
Changes Super.call() to always check the type of the first argument.

Fixes attribute lookup so that the lookup is on PyTDClass(super) when the
super instance is unbound. Based on the error message that the interpreter
gives for super(cls).__new__(cls) ('<cls> is not a subtype of super'), I
think the __new__ method being called here is super.__new__, which is why
I think the lookup needs to be on PyTDClass(super).

--
PiperOrigin-RevId: 159131784
MOS_MIGRATED_REVID=159131784