Don't allow formal types to be passed to functions.
When we encounter a function argument that is a formal type, log an
invalid_typevar error and change the argument to an unsolvable, so that
the matcher doesn't crash.

This covers the case of typevars being passed to typing.cast, so we no
longer need to specially check for that in typing.Cast.call().

Allows us to use setitem_slot instead of setitem for setting up a kwargs
dictionary, since we no longer have to worry about setitem_slot's call to
call_pytd crashing with an assertion error.

Also fixes a couple of lint warnings in test_typing.py.

--
PiperOrigin-RevId: 159259060
MOS_MIGRATED_REVID=159259060