Create a new CFG node after a STORE_FAST.
This fixes the "test_loop_and_if" test case in test_flow.py.
It prevents the solver from "single-stepping" over variable assignments inside
that loop.
Note that this triggers another bug (in tests/test_splits.py) if APPROXIMATE
is set to "True" in cfg.py. So we need to make strict-attr-checking the default
before applying this.

--
PiperOrigin-RevId: 159260865
MOS_MIGRATED_REVID=159260865