Make AliasingDict.add_alias() work properly when the dict is non-empty.
Since we've been using AliasingDict in combination with LazyDict, add_alias is
usually called before any of the lazy values have been evaluated, so
"assert alias not in self" in add_alias() always passes, even though the
condition is too strict. Modifies the assertion to allow the alias to be
present as long as it's a key in the alias map. Also fixes a lint warning
about special methods having the wrong signature by adding unused varargs and
kwargs to a bunch of unimplemented methods.

--
PiperOrigin-RevId: 159500240
MOS_MIGRATED_REVID=159500240