Don't allow setting the module on an abstract.Module.
I noticed while investigating testTemplateConstruction that sometimes the
full name of the builtins module was "__builtin__.__builtin__", which is
silly. Changes abstract.Module to always return None when its module attribute
is accessed and to ignore attempts to set its module (after making sure that
that the attempted new value is something reasonable).

--
PiperOrigin-RevId: 159876625
MOS_MIGRATED_REVID=159876625