Make FunctionType mutable.
The reason for the unreplaced NamedType in xml.etree.ElementTree is that
FillInModuleClasses fills in ClassType pointers, replacing old versions of
classes, but leaves FunctionType nodes unchanged, even though they can
similarly point to old versions of functions.

Makes FunctionType mutable so that the "function" attribute can be overwritten
by FillInModuleClasses (not done in this change).

--
PiperOrigin-RevId: 160313366
MOS_MIGRATED_REVID=160313366