Improve VirtualMachine.call_binary_operator.
Changes call_binary_operator to loop over the left and right side bindings so
that we can determine individually for each pair of values whether to call the
operator or the reverse operator. (Sometimes we have to try both, to emulate
the behavior of builtins that return NotImplemented.)

The reverse operator should be tried first when the right side is a subclass
of the left and overrides or newly defines a reverse operator. So for example,
in this case, Bar.__radd__ will be called:

class Foo(object):
  def __add__(self, other): ...
class Bar(Foo):
  def __radd__(self, other): ...
Foo() + Bar()  # Bar.__radd__ used

But in this case, since Bar doesn't touch __radd__, Foo.__add__ is called:

class Foo(object):
  def __add__(self, other): ...
  def __radd__(self, other): ...
class Bar(Foo): ...
Foo() + Bar()  # Foo.__add__ used

--
PiperOrigin-RevId: 161234012
MOS_MIGRATED_REVID=161234012