Call __init__ on return values instantiated during matching.
matcher.py uses instantiate() rather than init_class(), which in most cases is
correct, since it creates a lot of throwaway instances for matching. But
occasionally one of these instances ends up being the return value, so we
need to call __init__ on it.

Also defines a dummy call_init() in vm.py to make abstract_test and vm_test
happy, and fixes a bug in special_builtins that this change exposes. (We were
sometimes calling get_class_attribute() without first checking that the
abstract value was in fact an abstract.Class.)

--
PiperOrigin-RevId: 161735655
MOS_MIGRATED_REVID=161735655