Fix our logic for deciding when to type-check an abstract method.
We were depending on the is_attribute_of_class attribute of functions, but
this attribute isn't set properly when (1) the method's return type is a
namedtuple, or (2) when the skip_repeat_calls flag is set to False.

Also changes assertEquals to assertEqual in a few places to make lint happy.

--
PiperOrigin-RevId: 162526974
MOS_MIGRATED_REVID=162526974