Report an error for a definition like class B(A[T], A[int]): ...
Report an error when a type parameter in a class's template is also given a
concrete value.  Changes AliasingDict to allow turning a name with a value into
an alias, so that class B(A[int], A[T]): ... is an error rather than a crash.

Also renames some asserts and corrects a typo, at the direction of gpylint.

--
PiperOrigin-RevId: 164220880
MOS_MIGRATED_REVID=164220880