Fix node bug in converter.py.
When converting the parameters of a GenericType, we need to use the current
node, since they might be type parameters, referring to concrete
instances declared at a non-root node. We wouldn't be able to reach these
from the root node.
Also fixed a typo found by pylint.

--
PiperOrigin-RevId: 165190425
MOS_MIGRATED_REVID=165190425