Unresolved classes could escape from the pickling code.
This could happen if
A ClassType in module1, used for say a Signature, references something in module2.
Visitors.LookupExternalTypes could not resolve something and instead replaced the ClassType with AnythingType.
If a local reference in module1 pointed to Signature, the local reference would point to an unresolved ClassType instead of AnythingType.

--
PiperOrigin-RevId: 167182863
MOS_MIGRATED_REVID=167182863