Remove variable callbacks.
Computes a changestamp from the sum of the number of variables and the number
of variable bindings in a MonitorDict, so that the changestamp changes whenever
a variable or binding is added.

Fixes a test in test_tracebacks that, as far as I can tell, only worked
previously because we made some unnecessary change to the changestamp that
caused f() to be called twice with the same argument.

--
PiperOrigin-RevId: 168388414
MOS_MIGRATED_REVID=168388414