Make ObjectPredicate inherit from BuiltinFunction.
Fixes the crash in matcher.py, since the code that can't handle a bare
abstract.Function knows how to handle abstract.PyTDFunction.

Also moves BuiltinFunction, Abs, and Next earlier in the special_builtins file
to group them with the object predicate classes.

Now that isinstance is a PyTDFunction, pytype is able to infer a bit more about
it, and a couple of tests have been tweaked to reflect that.

--
PiperOrigin-RevId: 168721113
MOS_MIGRATED_REVID=168721113