Explain why a type didn't match a protocol.
This puts more information into error messages, explaining why we failed to
match a protocol. It only does this for parameters, and only for the top
level type. (So e.g. passing a List[X] argument for a List[Iterable] parameter
will *not* complain about a missing "__iter__" on X)

--
PiperOrigin-RevId: 169442089
MOS_MIGRATED_REVID=169442089