Fix a subtle import issue.
If a module (package) is called "foo.bar", and is implemented in
"foo/bar/__init__.py", then it can reference modules that *start* with
"foo.bar" (e.g. "foo.bar.baz", implemented in foo/bar/baz.py). Make sure that
when we rename the local module, we don't rename references to modules below
the current package. (In this example, do not rename "foo.bar.baz", but do
rename "foo.bar")


Also includes the following changes:

Convert properties to constants at the module rather than class level.

Lets us have a more precise notion of when a property type depends on a type
variable (in which case we have to leave it as a method to be resolved at use
time).

Also cleans up a lot of property/method splitting code in build_class.

--
PiperOrigin-RevId: 169739601
MOS_MIGRATED_REVID=169739601