Fix a bug caused by incorrectly reusing the seen set in output.py.
In the test case, because we set the list contents to the tuple twice, the
tuple is added to the seen set twice, and the second time, pytype mistakenly
thinks it's encountered a recursive definition. Copying `seen` fixes the
problem because the tuple is now added once to each copy.

--
PiperOrigin-RevId: 171987807
MOS_MIGRATED_REVID=171987807