This provides a code path around shlex.
Shlex has a large performance impact. For big imports_info files (>2k lines), this becomes one of the dominant sources of spend runtime.

This change provides a way for callers to provide a separator strings. Which is used to split key and values in the imports info file. This split can be done with string.split, which is free compared to the state machine used by shlex.

--
PiperOrigin-RevId: 172470484
MOS_MIGRATED_REVID=172470484