Don't analyze the function used to build a class.
To build a class, Python first creates and runs a function to set up all of
the locals, and then uses the last frame from that call to make the class.
pytype was analyzing that function in the loop in analyze_toplevel() that
goes through not-yet-analyzed interpreter functions. Class methods were
sometimes recreated as unbound functions as part of this analysis, leading
to weird behavior. Let's just stop analyzing these class-building functions.

In Python 2, we detect a class-building function by the presence of the
LOAD_LOCALS opcode. In Python 3, it's the function that BUILD_CLASS runs.

--
PiperOrigin-RevId: 175321657
MOS_MIGRATED_REVID=175321657